Added test for pipeline, add stderr channel as well to handle the errors and made pipeline work using intermediate buffers and connecting them through channel.
Test_cases are added for the pipeline but still tests are failing mainly due to string matching errors, foundation for pipeline has been created.
Stderr channel component for complete pipeline is also added.
We are able to get the output correct from the pipeline but still refactoring needs to be done.